# codigo-fonte/frontend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "IgescConecta.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/courses/search": {
      "post": {
        "tags": [
          "Courses"
        ],
        "operationId": "ListCourse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListCourseRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ListCourseRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ListCourseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ListCourseViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCourseViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCourseViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{id}": {
      "get": {
        "tags": [
          "Courses"
        ],
        "operationId": "GetCourseById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetCourseByIdViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCourseByIdViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCourseByIdViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{courseId}": {
      "put": {
        "tags": [
          "Courses"
        ],
        "operationId": "EditCourse",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditCourseRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EditCourseRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EditCourseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EditCourseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditCourseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditCourseResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Courses"
        ],
        "operationId": "DeleteCourse",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCourseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCourseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCourseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses": {
      "post": {
        "tags": [
          "Courses"
        ],
        "operationId": "CreateCourse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCourseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCourseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCourseResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projectdocuments/{id}": {
      "put": {
        "tags": [
          "ProjectDocuments"
        ],
        "operationId": "UpdateProjectDocument",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "File": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "File": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProjectDocumentResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProjectDocumentResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProjectDocumentResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ProjectDocuments"
        ],
        "operationId": "GetProjectDocument",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetProjectDocumentResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProjectDocumentResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProjectDocumentResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ProjectDocuments"
        ],
        "operationId": "DeleteProjectDocument",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/projectdocuments": {
      "get": {
        "tags": [
          "ProjectDocuments"
        ],
        "operationId": "ListProjectDocuments",
        "parameters": [
          {
            "name": "ProjectProgramId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ListProjectDocumentsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProjectDocumentsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProjectDocumentsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ProjectDocuments"
        ],
        "operationId": "CreateProjectDocument",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "ProjectProgramId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "Name": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "File": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "ProjectProgramId": {
                  "style": "form"
                },
                "Name": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "File": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProjectDocumentResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProjectDocumentResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProjectDocumentResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CourseViewModel": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "teamsCount": {
            "type": "integer",
            "format": "int32"
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "CreateCourseRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CreateCourseResponse": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CreateProjectDocumentResponse": {
        "type": "object",
        "properties": {
          "projectDocumentId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "DeleteCourseResponse": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "EditCourseRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "EditCourseResponse": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "Filter": {
        "type": "object",
        "properties": {
          "propertyName": {
            "type": "string"
          },
          "operation": {
            "$ref": "#/components/schemas/Op"
          },
          "value": { }
        },
        "additionalProperties": false
      },
...
      "GetCourseByIdViewModel": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "teamsCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "createdBy": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "integer",
            "format": "int32"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "teams": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamViewModel"
            }
          }
        },
        "additionalProperties": false
      },
...
      "GetProjectDocumentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "projectProgramId": {
            "type": "integer",
            "format": "int32"
          },
          "hasFile": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ListCourseRequest": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Filter"
            }
          },
          "statusFilter": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ListCourseViewModel": {
        "type": "object",
        "properties": {
          "totalItems": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseViewModel"
            }
          }
        },
        "additionalProperties": false
      },
...
      "ListProjectDocumentsResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectDocumentListItem"
            }
          }
        },
        "additionalProperties": false
      },
...
      "Op": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11,
          12,
          13,
          14,
          15
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ProjectDocumentListItem": {
        "type": "object",
        "properties": {
          "projectDocumentId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "TeamViewModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "lessonTime": {
            "type": "string",
            "nullable": true
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "finish": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateProjectDocumentResponse": {
        "type": "object",
        "properties": {
          "projectDocumentId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Informe: Bearer {seu_token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}